requirements.txt
flask
mysqlclient
python-dotenv
pyTelegramBotAPI
openai
peewee
